Added nutrient extractor

+++ include/models/usda/Nutrient.h
@@ -7,7 +7,7 @@ typedef struct {
   int id;
   std::string name;
   std::string unit_name;
-  int nutrient_nbr;
+  float nutrient_nbr;
   float rank;
 } Nutrient;
 } // namespace USDA
